feat/split-e2e-tests: splitted more tests
