+++ apps/chat-e2e/src/tests/sharedWithMeFolders.test.ts
@@ -15,8 +15,7 @@ import {
   MenuOptions,
   ModelIds,
 } from '@/src/testData';
-import { keys } from '@/src/ui/keyboard';
-import { GeneratorUtil, ItemUtil, ModelsUtil } from '@/src/utils';
+import { ModelsUtil } from '@/src/utils';
 import { expect } from '@playwright/test';
 
 const nestedLevel = 3;
@@ -25,103 +24,6 @@ let defaultModel: DialAIEntityModel;
 dialTest.beforeAll(async () => {
   defaultModel = ModelsUtil.getDefaultModel()!;
 });
-
-dialSharedWithMeTest(
-  'Shared with me. Share single chat in Folder.\n' +
-    'Shared chat disappears from Shared with me if the original was renamed.\n' +
-    'Shared with me. Structure appears only once if to open the same link several times',
-  async ({
-    dialHomePage,
-    folderConversations,
-    conversationDropdownMenu,
-    conversations,
-    confirmationDialog,
-    page,
-    localStorageManager,
-    additionalShareUserDialHomePage,
-    additionalShareUserSharedWithMeConversations,
-    conversationData,
-    dataInjector,
-    mainUserShareApiHelper,
-    additionalUserShareApiHelper,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-1827', 'EPMRTC-2773', 'EPMRTC-1854');
-    let conversationInFolder: FolderConversation;
-    let conversation: Conversation;
-    let shareByLinkResponse: ShareByLinkResponseModel;
-
-    await dialSharedWithMeTest.step(
-      'Prepare shared conversation inside folder',
-      async () => {
-        conversationInFolder =
-          conversationData.prepareDefaultConversationInFolder();
-        await dataInjector.createConversations(
-          conversationInFolder.conversations,
-          conversationInFolder.folders,
-        );
-        conversation = conversationInFolder.conversations[0];
-        await localStorageManager.setSelectedConversation(conversation);
-        shareByLinkResponse = await mainUserShareApiHelper.shareEntityByLink([
-          conversation,
-        ]);
-      },
-    );
-
-    await dialSharedWithMeTest.step(
-      'Open twice share link by another user and verify only chat stays under Shared with section',
-      async () => {
-        for (let i = 1; i <= 2; i++) {
-          await additionalShareUserDialHomePage.openHomePage(
-            { iconsToBeLoaded: [defaultModel!.iconUrl] },
-            ExpectedConstants.sharedConversationUrl(
-              shareByLinkResponse.invitationLink,
-            ),
-          );
-          await additionalShareUserDialHomePage.waitForPageLoaded();
-          await additionalShareUserSharedWithMeConversations
-            .getConversationByName(conversation.name)
-            .waitFor();
-        }
-      },
-    );
-
-    await dialSharedWithMeTest.step(
-      'Rename shared chat name and verify renamed chat is not shared any more',
-      async () => {
-        const updatedName = GeneratorUtil.randomString(7);
-        await dialHomePage.openHomePage({
-          iconsToBeLoaded: [defaultModel!.iconUrl],
-        });
-        await dialHomePage.waitForPageLoaded();
-        await folderConversations.openFolderEntityDropdownMenu(
-          conversationInFolder.folders.name,
-          conversation.name,
-        );
-        await conversationDropdownMenu.selectMenuOption(MenuOptions.rename);
-        await conversations.openEditConversationNameMode(
-          conversation.name,
-          updatedName,
-        );
-        await page.keyboard.press(keys.enter);
-        if (await confirmationDialog.isVisible()) {
-          await confirmationDialog.confirm({ triggeredHttpMethod: 'DELETE' });
-        }
-
-        const sharedEntities =
-          await additionalUserShareApiHelper.listSharedWithMeEntities();
-        expect
-          .soft(
-            sharedEntities.resources.find(
-              (e) => e.name === updatedName || e.name === conversation.name,
-            ),
-            ExpectedMessages.conversationIsNotShared,
-          )
-          .toBeUndefined();
-      },
-    );
-  },
-);
 
 dialSharedWithMeTest(
   'Shared with me. Share root Folder.\n' +
@@ -258,327 +160,6 @@ dialSharedWithMeTest(
             .getFolderEntity(nestedFolders[i].name, nestedConversations[i].name)
             .waitFor({ state: 'hidden' });
         }
-      },
-    );
-  },
-);
-
-dialSharedWithMeTest(
-  'Shared with me. Share Folder in the middle.\n' +
-    'Shared folder disappears from Shared with me if the original was renamed',
-  async ({
-    dialHomePage,
-    page,
-    folderConversations,
-    folderDropdownMenu,
-    confirmationDialog,
-    localStorageManager,
-    additionalShareUserDialHomePage,
-    additionalShareUserSharedFolderConversations,
-    conversationData,
-    dataInjector,
-    mainUserShareApiHelper,
-    additionalShareUserLocalStorageManager,
-    additionalUserShareApiHelper,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-1829', 'EPMRTC-2771');
-    let nestedFolders: FolderInterface[];
-    let nestedConversations: Conversation[];
-    let shareByLinkResponse: ShareByLinkResponseModel;
-
-    await dialSharedWithMeTest.step(
-      'Prepare conversations inside nested folder structure, share middle level folder',
-      async () => {
-        nestedFolders = conversationData.prepareNestedFolder(nestedLevel);
-        conversationData.resetData();
-        nestedConversations =
-          conversationData.prepareConversationsForNestedFolders(nestedFolders);
-        await dataInjector.createConversations(
-          nestedConversations,
-          ...nestedFolders,
-        );
-        shareByLinkResponse = await mainUserShareApiHelper.shareEntityByLink(
-          [nestedConversations[nestedLevel - 1]],
-          true,
-        );
-      },
-    );
-
-    await dialSharedWithMeTest.step(
-      'Open share link by another user and verify the structure below shared folder is displayed under Shared with section',
-      async () => {
-        await additionalShareUserLocalStorageManager.setSelectedConversation(
-          nestedConversations[nestedLevel],
-        );
-        await additionalShareUserDialHomePage.openHomePage(
-          { iconsToBeLoaded: [defaultModel!.iconUrl] },
-          ExpectedConstants.sharedConversationUrl(
-            shareByLinkResponse.invitationLink,
-          ),
-        );
-        await additionalShareUserDialHomePage.waitForPageLoaded();
-        for (let i = nestedLevel - 1; i <= nestedLevel; i++) {
-          await additionalShareUserSharedFolderConversations
-            .getFolderEntity(nestedFolders[i].name, nestedConversations[i].name)
-            .waitFor();
-        }
-        for (let i = 0; i < nestedLevel - 1; i++) {
-          await additionalShareUserSharedFolderConversations
-            .getFolderEntity(nestedFolders[i].name, nestedConversations[i].name)
-            .waitFor({ state: 'hidden' });
-        }
-      },
-    );
-
-    await dialSharedWithMeTest.step(
-      'Rename shared folder and verify it is not displayed under Shared with section',
-      async () => {
-        await localStorageManager.setSelectedConversation(
-          nestedConversations[nestedLevel],
-        );
-        const updatedFolderName = GeneratorUtil.randomString(7);
-        await dialHomePage.openHomePage();
-        await dialHomePage.waitForPageLoaded();
-        await folderConversations.openFolderDropdownMenu(
-          nestedFolders[nestedLevel - 1].name,
-        );
-        await folderDropdownMenu.selectMenuOption(MenuOptions.rename);
-
-        await folderConversations.editFolderName(
-          nestedFolders[nestedLevel - 1].name,
-          updatedFolderName,
-        );
-        await page.keyboard.press(keys.enter);
-        if (await confirmationDialog.isVisible()) {
-          await confirmationDialog.confirm({ triggeredHttpMethod: 'POST' });
-        }
-
-        const sharedEntities =
-          await additionalUserShareApiHelper.listSharedWithMeEntities();
-        expect
-          .soft(
-            sharedEntities.resources.find(
-              (e) =>
-                e.name === updatedFolderName ||
-                e.name === nestedConversations[nestedLevel - 1].name,
-            ),
-            ExpectedMessages.folderIsNotShared,
-          )
-          .toBeUndefined();
-      },
-    );
-  },
-);
-
-dialSharedWithMeTest(
-  'Shared with me. Folder structure is updated if to add new folder with chat to original folder',
-  async ({
-    dialHomePage,
-    folderConversations,
-    localStorageManager,
-    conversationData,
-    dataInjector,
-    mainUserShareApiHelper,
-    additionalUserShareApiHelper,
-    additionalUserItemApiHelper,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-2758');
-    let sharedConversationInFolder: FolderConversation;
-    let conversationInFolder: FolderConversation;
-    let singleConversation: Conversation;
-    let sharedConversation: Conversation;
-    let movedConversationInFolder: Conversation;
-    let shareByLinkResponse: ShareByLinkResponseModel;
-    let sharedFolderName: string;
-    let folderName: string;
-
-    await dialSharedWithMeTest.step(
-      'Prepare shared conversation inside folder, one more conversation inside folder and single conversation',
-      async () => {
-        sharedConversationInFolder =
-          conversationData.prepareDefaultConversationInFolder();
-        sharedFolderName = sharedConversationInFolder.folders.name;
-        conversationData.resetData();
-        conversationInFolder =
-          conversationData.prepareDefaultConversationInFolder();
-        folderName = conversationInFolder.folders.name;
-        conversationData.resetData();
-        singleConversation = conversationData.prepareDefaultConversation();
-
-        await dataInjector.createConversations(
-          [
-            ...sharedConversationInFolder.conversations,
-            ...conversationInFolder.conversations,
-            singleConversation,
-          ],
-          sharedConversationInFolder.folders,
-          conversationInFolder.folders,
-        );
-        sharedConversation = sharedConversationInFolder.conversations[0];
-        shareByLinkResponse = await mainUserShareApiHelper.shareEntityByLink(
-          [sharedConversation],
-          true,
-        );
-        await additionalUserShareApiHelper.acceptInvite(shareByLinkResponse);
-      },
-    );
-
-    await dialSharedWithMeTest.step(
-      'Move conversation inside folder and single conversation inside shared folder',
-      async () => {
-        movedConversationInFolder = conversationInFolder.conversations[0];
-        movedConversationInFolder.id = movedConversationInFolder.id.replace(
-          folderName,
-          `${sharedFolderName}/${folderName}`,
-        );
-        movedConversationInFolder.folderId =
-          movedConversationInFolder.folderId.replace(
-            folderName,
-            `${sharedFolderName}/${folderName}`,
-          );
-        const singleConversationName = `${singleConversation.model.id}${ItemUtil.conversationIdSeparator}${singleConversation.name}`;
-        singleConversation.id = singleConversation.id.replace(
-          singleConversationName,
-          `${sharedFolderName}/${singleConversationName}`,
-        );
-        singleConversation.folderId =
-          singleConversation.folderId + `/${sharedFolderName}`;
-
-        await dataInjector.updateConversations([
-          movedConversationInFolder,
-          singleConversation,
-        ]);
-      },
-    );
-
-    await dialSharedWithMeTest.step(
-      'Open app by main user and verify moved folder does not have shared icon',
-      async () => {
-        await localStorageManager.setSelectedConversation(sharedConversation);
-        await dialHomePage.openHomePage();
-        await dialHomePage.waitForPageLoaded();
-        await folderConversations
-          .getFolderArrowIcon(folderName)
-          .waitFor({ state: 'hidden' });
-      },
-    );
-
-    await dialSharedWithMeTest.step(
-      'Verify moved folder with content and single conversation are shared for the user',
-      async () => {
-        const sharedEntities =
-          await additionalUserShareApiHelper.listSharedWithMeEntities();
-        expect
-          .soft(
-            sharedEntities.resources.find((e) => e.name === sharedFolderName),
-            ExpectedMessages.folderIsShared,
-          )
-          .toBeDefined();
-
-        const sharedWithMeItems: BackendDataEntity[] = [];
-        for (const sharedEntity of sharedEntities.resources) {
-          const sharedItems = await additionalUserItemApiHelper.listItem(
-            sharedEntity.url,
-          );
-          sharedWithMeItems.push(...sharedItems);
-        }
-        expect
-          .soft(
-            sharedWithMeItems.find(
-              (i) => i.url === movedConversationInFolder.id,
-            ),
-            ExpectedMessages.conversationIsShared,
-          )
-          .toBeDefined();
-        expect
-          .soft(
-            sharedWithMeItems.find((i) => i.url === singleConversation.id),
-            ExpectedMessages.conversationIsShared,
-          )
-          .toBeDefined();
-      },
-    );
-  },
-);
-
-dialSharedWithMeTest(
-  'Shared with me. Folder structure is updated if to remove chat from original folder',
-  async ({
-    conversationData,
-    dataInjector,
-    mainUserShareApiHelper,
-    additionalUserShareApiHelper,
-    additionalUserItemApiHelper,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-2759');
-    let sharedConversationInFolder: FolderConversation;
-    let sharedConversation: Conversation;
-    let shareByLinkResponse: ShareByLinkResponseModel;
-    let sharedFolderName: string;
-
-    await dialSharedWithMeTest.step(
-      'Prepare shared conversation inside folder',
-      async () => {
-        sharedConversationInFolder =
-          conversationData.prepareDefaultConversationInFolder();
-        sharedFolderName = sharedConversationInFolder.folders.name;
-
-        await dataInjector.createConversations(
-          sharedConversationInFolder.conversations,
-          sharedConversationInFolder.folders,
-        );
-        sharedConversation = sharedConversationInFolder.conversations[0];
-        shareByLinkResponse = await mainUserShareApiHelper.shareEntityByLink(
-          [sharedConversation],
-          true,
-        );
-        await additionalUserShareApiHelper.acceptInvite(shareByLinkResponse);
-      },
-    );
-
-    await dialSharedWithMeTest.step(
-      'Move shared conversation out of folder',
-      async () => {
-        sharedConversation.id = sharedConversation.id.replace(
-          `/${sharedFolderName}`,
-          '',
-        );
-        sharedConversation.folderId = sharedConversation.folderId.replace(
-          `/${sharedFolderName}`,
-          '',
-        );
-        await dataInjector.updateConversations([sharedConversation]);
-      },
-    );
-
-    await dialSharedWithMeTest.step(
-      'Verify only folder is shared with user',
-      async () => {
-        const sharedEntities =
-          await additionalUserShareApiHelper.listSharedWithMeEntities();
-        expect
-          .soft(
-            sharedEntities.resources.find((e) => e.name === sharedFolderName),
-            ExpectedMessages.folderIsShared,
-          )
-          .toBeDefined();
-
-        const sharedWithMeItems: BackendDataEntity[] = [];
-        for (const sharedEntity of sharedEntities.resources) {
-          const sharedItems = await additionalUserItemApiHelper.listItem(
-            sharedEntity.url,
-          );
-          sharedWithMeItems.push(...sharedItems);
-        }
-        expect
-          .soft(
-            sharedWithMeItems.find((i) => i.url === sharedConversation.id),
-            ExpectedMessages.conversationIsNotShared,
-          )
-          .toBeUndefined();
       },
     );
   },
@@ -775,87 +356,6 @@ dialSharedWithMeTest(
         expect
           .soft(isCodeContentVisible, ExpectedMessages.codeIsVisibleInResponse)
           .toBeTruthy();
-      },
-    );
-  },
-);
-
-dialSharedWithMeTest(
-  'Shared folder disappears from Shared with me if the original was deleted.\n' +
-    'Shared chat disappears from Shared with me if the original was deleted',
-  async ({
-    conversationData,
-    dataInjector,
-    mainUserShareApiHelper,
-    additionalUserShareApiHelper,
-    itemApiHelper,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-2770', 'EPMRTC-2772');
-    let conversationInFolder: FolderConversation;
-    let conversation: Conversation;
-    let shareByLinkFolderResponse: ShareByLinkResponseModel;
-    let shareByLinkConversationResponse: ShareByLinkResponseModel;
-
-    await dialSharedWithMeTest.step(
-      'Prepare shared folder with conversation and single shared conversation',
-      async () => {
-        conversationInFolder =
-          conversationData.prepareDefaultConversationInFolder();
-        conversationData.resetData();
-        conversation = conversationData.prepareDefaultConversation();
-        await dataInjector.createConversations(
-          [conversation, ...conversationInFolder.conversations],
-          conversationInFolder.folders,
-        );
-
-        shareByLinkFolderResponse =
-          await mainUserShareApiHelper.shareEntityByLink(
-            conversationInFolder.conversations,
-            true,
-          );
-        await additionalUserShareApiHelper.acceptInvite(
-          shareByLinkFolderResponse,
-        );
-
-        shareByLinkConversationResponse =
-          await mainUserShareApiHelper.shareEntityByLink([conversation]);
-        await additionalUserShareApiHelper.acceptInvite(
-          shareByLinkConversationResponse,
-        );
-      },
-    );
-
-    await dialTest.step(
-      'Delete shared folder and conversation by main user',
-      async () => {
-        await itemApiHelper.deleteConversation(
-          conversationInFolder.conversations[0],
-        );
-        await itemApiHelper.deleteConversation(conversation);
-      },
-    );
-
-    await dialSharedWithMeTest.step(
-      'Verify folder and conversation are not shared with user any more',
-      async () => {
-        const sharedEntities =
-          await additionalUserShareApiHelper.listSharedWithMeEntities();
-        //TODO: enable when https://github.com/epam/ai-dial-chat/issues/1139 is fixed
-        // expect
-        //   .soft(
-        //     sharedEntities.resources.find(
-        //       (f) => f.name === conversationInFolder.folders.name,
-        //     ),
-        //     ExpectedMessages.folderIsNotShared,
-        //   )
-        //   .toBeUndefined();
-        expect
-          .soft(
-            sharedEntities.resources.find((c) => c.url === conversation.id),
-            ExpectedMessages.conversationIsNotShared,
-          )
-          .toBeUndefined();
       },
     );
   },
